backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
authlib
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
alembic
